Initial websockets implementation
The whole game is now served by node.js, and unfortunately requires a build
step to work (see README.md). It is also rendered by plain canvas instead
of PIXI, and no longer requires jQuery or Bootstrap.

The board model has changed for serialization purposes, and so has the
coordinate system- edges and vertices now have a single canonical
location and are no longer stored as a "sea of nodes."

Some logic has been left behind (mostly related to player resources and
building costs) and needs to be reimplemented on the server side. Good
thing we have Git, right?