Adding city functionality
Adding the ability for a player to build a city. Cities can only be built
onto an existing town. In order to do this, I had to change the type of
buildings[y][x][d] from being an integer to an object that contains the
player ID (INT) and the building type (INT).